requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
python-dotenv==1.2.1
pydantic==2.12.5
aiofiles==25.1.0
langchain==1.2.7
PyMuPDF==1.26.7
langchain-community==0.4.1
motor==3.7.1
openai==2.21.0
cohere==5.20.5
groq==1.0.0
sentence-transformers==5.2.3
# qdrant-client==1.17.0
qdrant-client==1.10.1
prometheus-client==0.24.1
starlette-exporter==0.23.0
fastapi-health==0.4.0